working with toolbar
